Completely refactored the command system
The command system I was building had some issues that were solved by
simply using the event manager on the bot. The new command system is
completely event based meaning that multiple things can hook on a
command and the entire thing runs in its own goroutine.

Some locks may be needed in some places because of that, to be seen.