Made sure to use a pointer for game.ValidateConfig
ValidateConfig compiles formats in that config to check them for errors,
UpdateFromConf then uses those validated formats as the configs on the
game. A pointer was previously unused which therefore made the config be
copied when its being validated, this meant that the formats added to
the game by UpdateFromConf were not compiled, causing a runtime error.