Fixed Index Out Of Range panic in On* callbacks
Some commands (namely kick, quit and part) have optional messages.
Originally it was assumed that these messages would always be filled.
This has proven to be not the case. The "expected" behaviour is now
implemented, that is, if the message does **not** exist, the message is
passed as an empty string.