Added sendToLocalGame bool on game regexps
This bool makes the result of the formatter attached to the regexp be
sent to the local game as well. The line is sent over the
SendLineFromOtherGame interface method, with the originating game sent
as the local game.

On the other side, SendLineFromOtherGame's implementation in game.Game
has been updated to check if the remote game is the same as the local
game, and if so, set the name to be given to the formatter to the string
LOCAL.

Also included in this commit is an updated XSD to reflext the added
attribute on the config side.

Closes #30