Made game.Manager the frontend struct to be run
This makes various changes to the game.Manager struct, such as adding a
command.Manager instance, a done condition, and a restarting bool
fields, and a Run and runBots method.

Run is similar to bot.Bot.Run(), it starts the bot and connects to
(currently) IRC, and starts games that are marked for autostart.

runBots uses the interfaces.Bot on Manager to continually restart bot
until such a time as the Manager is being shutdown