Added ConnConfig
ConnConfig holds the configuration for a given connection. The
configuration is held as a string on ConnConfig, and is parsed later by
the interface.Bot implementation during its construction. The added
UnmarshalXML and util functions are used to reconstruct the config from
a set of XML tokens.

This method was chosen over using ",innerxml" on a string object as it
allows for the preservation of attributes on the root tag, whereas
innerxml only allows access to the tags within the root tag. For bots
this was deemed unacceptable.