Fixed data race in AttachMultiple and its users
This adds an attachDirect method on Manager that sets the entry in the
event map to a given ID, rather than generating the ID on the fly. This
solves a race condition in AttachMultiple and co where the race
dectector saw reads and writes to the locally scoped (to the closure and
surrounding method) id variable. The id variable was set by a call to an
Attach* method, which returned the attached ID. Now, the ID is generated
first and not modified after initial generation. This is still safe as
nextID() is protected by a mutex