Removed prefix check from game.shouldBridge
The '#' prefix check in this method was an IRC specific check, which has
been superceded by the existence of HookMessage and HookPrivateMessage
on interface.Bot, which does the same check (for the irc implmentation).

As such for calls to shouldBridge, that check is always true, as none of
the On* methods on Game are called with private messages.