Finished IRC's impl of format.Transformer
This adds an IRC colour parser for mirc style IRC colour codes, and uses
it to transform IRC lines to format/transformer.Transformer's
intermediate format.

Note that this does NOT support IRCCloud style hex colours as of yet. If
support is added, it will only be parsing and not emitting.