Changed permissions of bot control commands
Previously both game and bot control fell under the same level, meaning
that if someone controls the games they can control the bot, which I
feel is questionable. This change also makes the raw command on games a
level three, as it (depending on the game) allows for system access. And
lastly status requires 0 perms, because users should be able to ask the
bot for the status of their games.

Closes #38