Moved CTCP ACTION support to IRC
Previously, game.OnMessage was parsing lines for CTCP, which does not
make sense in an abstracted world.

This makes interfaces.game's OnMessage callback have an additional bool
noting whether or not the message is an "action" message. How each bot
understands an "Action" is up to them. They can simply choose to never
have actions.

Additionally, the Hooker interface's HookMessage has been updated with a
similar IsAction bool, to indicate to the callback that the message is
an action.

The implementations of these interfaces has also been updated in this
commit.