Added workaround for weird MineCraft behaviour
MineCraft JSON formatting has a quirk that when you use an array of
objects contraining strings, the "default" formatting setup is done by
the first string. This means that, for example, if you were to change
the colour and italicise a prefix for a message, the entire message
would be coloured and italicised unless they're explicitly turned off.
This commit solves this issue by emitting an empty string with no
formatting as the first item in the array if the "original" first item
had any sort of formatting.