Abstracted ConfigHolder, added Transport configs
Various constructs in the config system are re-parsed later after
selection of an Interface implementation is selected, such as for bots
or for transports. Before these configs all used their own structs with
their own (subtly different) implementations of unmarshaler. This commit
changes them to use the same base implementation.