Added network transport
This commit adds an (unfinished, but relatively working) implementation
of a network based transport for use with games.

The network transport comes in two parts, first is an in-GGGB
clientside, which talks to the server, currently called prog by way of
some network socket. All testing has taken place using unix sockets, but
network sockets should work in the same way. The communication format is
go GOBs, by way of the go RPC lib. This is subject to change as the RPC
lib is feature frozen and has some things that I don't particularly like.