Fixed duplicated working directory configs
When it was created, the network.Config struct did not embed baseConfig.
Instead, it had its own working directory field. This caused a bug where
it could appear in the config that the working directory was set, but
the code used the other struct field, causing the code-side working
directory to be inferred based on the config location.