Fixed lost lines on processes that exit quickly
Okay this is a big one. So.

Originally. I was using stdioPipe methods for accessing stdio on
processes. This makes sense because, well, it does. What I missed in the
docs was that if you use the pipe types that you should finish all I/O
ops before using Wait(), which the code previously did not do. Now for
the most part this was fine as we didnt work with very quick exits when
running minecraft servers and such. But while I was testing I was
constantly plagued by that issue (eg using a python script to output a
bunch of lines very quickly)

This commit finally fixes this by ensuring that process.Wait (as in our
process, not the stdlib one) waits on IO finishing BEFORE calling Wait
on the process itself.

To do that there is a new file and function, which creates a pipe based
I/O watcher at the process layer. Using io.copy keeps the stream exactly
as it was, but lets us snoop and block until the stdio EOFs, and then we
close the other side of the pipe. Once both sides are closed we Done() a
waitgroup, which is watched by the Wait method on our Process struct.
Once everything is okay, the Wait method calls Wait on the process
(which should complete instantly but doesnt have to), which then finally
reaps the process and gets the exit code etc.

This does carry the possible goroutine leak of the io.copy goroutine,
but it SHOULD clean itself up, unless we're reset without being used,
which should never happen.