Switched to using a single method for stdio
Previously there would be two goroutines sitting repeatedly calling
GetStdLines with a bool to specify whether or not to do stdout etc. This
was wasteful and has been removed

Now we use one method, and each line relayed indicates whether it is
from stdout or stderr.

Additionally, instead of assuming all lines are unique, we use a
sequence number to ensure that all lines are dealt with only once. There
may still be some bugs surrounding this in edge cases.