Added IsCommandPrefix, renamed CommandPrefixes on Bot
This commit adds a method to Bot called IsCommandPrefix, it is used by
command.Manager to check for dynamic command prefixes that can be
changed at any time.

It also renames CommandPrefixes to StaticCommandPrefix (pending possible
removal), to better indicate that the results from that cannot be
removed after the fact.