Refactored comparisons to fix DeepEqual issues
The standard method of reflect.DeepEqual will not work for the Transport
config, as the "expected" types are created using a method that does not
set ALL values on the expected Tree struct, now the only values that are
not set are internal to the toml lib, they tell it about where in a file
this tree points to. As such they are not relevant to us, and can be
safely ignored when it comes to verifying that a config is as expected.

For this reason I have created a manually verifying if, which compares
most of the struct normally but only compares the Type of a transport
and the String() (which is a recreated TOML file) of the Tree