Fixed edge case causing hang on connect
While connecting, irc.Connect is called from irc.Run, if the perfect
storm of timing occurs, Connect will progress to the point where its
waiting for a 001 from the server, but the socket has been disconnected
or otherwise interrupted. Before this commit, the Connect method would
hang indefinitely waiting, to the point that even the `reconnect`
command would not release the hang.

irc.Connect now listens for both 001 AND a socket close event