Increased wait time before forced socket close
This simply increases the wait time before force closing the socket when
using irc.Disconnect. Previously the time was only a few milliseconds,
which is not enough time for a connection to clean up when the ping can
be upwards of half a second.

Additionally, if a forced socket closure occurs, it is logged as a WARN
message.